Fix concurrency exception for MR_inContext if receiver has temporary objectID
If `MR_inContext` is called on an `NSManagedObject` that has not yet
been saved, `MagicalRecord` first calls
`-obtainPermanentIDsForObjects:error:` on the `NSManagedObjectContext`
of the temporary object. However, if the two contexts are on different
queues this will result in a concurrency exception (if concurrency
debug is turned on in the scheme).